requirements.txt
openai~=0.27.0
aiogram~=2.25.1
python-dotenv~=1.0.0
googletrans~=4.0.0rc1